+++ src/main/java/uk/co/ramp/covid/simulation/population/Adult.java
@@ -2,6 +2,7 @@ package uk.co.ramp.covid.simulation.population;
 
 import uk.co.ramp.covid.simulation.covid.CovidParameters;
 import uk.co.ramp.covid.simulation.DailyStats;
+import uk.co.ramp.covid.simulation.parameters.PopulationParameters;
 import uk.co.ramp.covid.simulation.place.*;
 import uk.co.ramp.covid.simulation.util.ProbabilityDistribution;
