+++ src/main/java/uk/co/ramp/covid/simulation/population/Infant.java
@@ -2,6 +2,7 @@ package uk.co.ramp.covid.simulation.population;
 
 import uk.co.ramp.covid.simulation.covid.CovidParameters;
 import uk.co.ramp.covid.simulation.DailyStats;
+import uk.co.ramp.covid.simulation.parameters.PopulationParameters;
 
 public class Infant extends Person {
