+++ src/test/java/uk/co/ramp/covid/simulation/ExampleParametersTest.java
@@ -6,7 +6,7 @@ import org.junit.Test;
 
 import uk.co.ramp.covid.simulation.covid.CovidParameters;
 import uk.co.ramp.covid.simulation.io.ParameterReader;
-import uk.co.ramp.covid.simulation.population.PopulationParameters;
+import uk.co.ramp.covid.simulation.parameters.PopulationParameters;
 
 public class ExampleParametersTest {
