+++ src/test/java/uk/co/ramp/covid/simulation/ModelTest.java
@@ -7,7 +7,7 @@ import org.junit.Test;
 import com.google.gson.JsonParseException;
 
 import uk.co.ramp.covid.simulation.covid.CovidParameters;
-import uk.co.ramp.covid.simulation.population.PopulationParameters;
+import uk.co.ramp.covid.simulation.parameters.PopulationParameters;
 import uk.co.ramp.covid.simulation.util.SimulationTest;
 
 import java.io.IOException;
