+++ src/test/java/uk/co/ramp/covid/simulation/io/ParameterReaderTest.java
@@ -3,7 +3,7 @@ package uk.co.ramp.covid.simulation.io;
 import org.junit.After;
 import org.junit.Test;
 import uk.co.ramp.covid.simulation.covid.CovidParameters;
-import uk.co.ramp.covid.simulation.population.PopulationParameters;
+import uk.co.ramp.covid.simulation.parameters.PopulationParameters;
 import java.io.IOException;
 
 import static org.junit.Assert.*;
