+++ src/test/java/uk/co/ramp/covid/simulation/place/HospitalTest.java
@@ -6,6 +6,7 @@ import com.google.gson.JsonParseException;
 
 import uk.co.ramp.covid.simulation.DailyStats;
 import uk.co.ramp.covid.simulation.Time;
+import uk.co.ramp.covid.simulation.parameters.PopulationParameters;
 import uk.co.ramp.covid.simulation.population.*;
 import uk.co.ramp.covid.simulation.util.SimulationTest;
