+++ src/test/java/uk/co/ramp/covid/simulation/population/PopulationTest.java
@@ -3,6 +3,7 @@ package uk.co.ramp.covid.simulation.population;
 import org.junit.Before;
 import org.junit.Test;
 import uk.co.ramp.covid.simulation.DailyStats;
+import uk.co.ramp.covid.simulation.parameters.PopulationParameters;
 import uk.co.ramp.covid.simulation.place.*;
 import uk.co.ramp.covid.simulation.place.householdtypes.*;
 import uk.co.ramp.covid.simulation.testutil.PopulationGenerator;
